Fix Graph xAxis Bug

+++ app/src/main/java/co/specialforce/data/response/getWeight/WeightArray.kt
@@ -1,3 +1,3 @@
 package co.specialforce.data.response.getWeight
 
-data class WeightArray(val profile_id: Int, val date: String, val weight: Int)
+data class WeightArray(val profile_id: Int, val date: String, val weight: Float)
